splitting ndes.py into ndes_pt.py and ndes_tf.py
